split into server and client components, adds Suspense
